requirements.txt
streamlit
requests
jamaibase
flask
pycountry
fastapi
python-dotenv
supabase